updates to ansible mjl-user playbook and readme
Note that before adding the append: yes and sudo to the list of groups my mjl user was *removed* from the sudo group and then I no longer had any admin access at all. I restored it by booting to a different OS and mounting the drive and putting my user back in the sudo group.