feat(ansible): add to apt pref file firefox-no-snap to prefer mozilla repo if present
This is done by adding a section to set the priority of the mozillateam PPA to 100, so it will install if no higher priority repo with firefox exists, but otherwise the higher priority one (mozilla w/ its default priority) will win out.

I'm doing this because the PPA version string for basically the same version was "higher" so it would replace the mozilla version a day or 2 later.