feat(ansible): yet more pi-client playbook changes as I try to create the pi-eliza box
the client name must now be specified on the commandline

I added a task to set the iptables to nat the wired interface subnet when forwarding it to the wifi interface. I'm not sure why, but this may be what was needed to get it working properly when the vpn is up.